backend/requirements.txt
asgiref==3.5.0
Django==3.2.9
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djongo==1.3.6
dnspython==2.2.1
dongo==0.4.0
PyJWT==1.7.1
pymongo==3.12.3
pytz==2021.3
six==1.16.0
sqlparse==0.2.4
tzdata==2021.5
